Develop

+++ src/entities/User/useCases/DeleteUser/DeleteUserUseCase.ts
@@ -1,5 +1,6 @@
 import { IUserRepository } from "../../repositories/interfaces/IUserRepository";
 import GetUserUseCase from "../GetUser/GetUserUseCase";
+import { NotFoundError } from "../../../../helpers/api-erros";
 
 export default class DeleteUserUseCase {
     private userRepository: IUserRepository;
@@ -11,7 +12,10 @@ export default class DeleteUserUseCase {
     }
 
     async execute(Id: number) {
-        const user = await this.getUserUseCase.execute(Id);
+        const user = await this.userRepository.findById(Id);
+        if (!user) {
+            throw new NotFoundError("Usuário não encontrado. Id: " + Id);
+        }
         return this.userRepository.delete(Id); 
     }
 }
